Refactor: Organize API calls into dedicated service folder
- Created  to centralize API calls
- Added separate files for images, recipe and user- Moved all API calls related to uploading and deleting images (recipe images & avatars) into
- Updated imports to use the new structure

#54
