+++ src/service/api/recipe.service.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+const api = axios.create({
+  baseURL: import.meta.env.VITE_API_URL || "http://localhost:5005",
+});
